simplifications on orderings
