+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/combinatorics/TransitiveClosure.java
@@ -1,9 +1,9 @@
 /*
- * TopologicalSort.java
+ * TransitiveClosure.java
  *
  * This source file is part of the FoundationDB open source project
  *
- * Copyright 2015-2020 Apple Inc. and the FoundationDB project authors
+ * Copyright 2015-2021 Apple Inc. and the FoundationDB project authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,7 +18,7 @@
  * limitations under the License.
  */
 
-package com.apple.foundationdb.record.query.plan.temp;
+package com.apple.foundationdb.record.query.combinatorics;
 
 import com.apple.foundationdb.annotation.API;
 import com.google.common.base.Preconditions;
