+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/temp/Quantifiers.java
@@ -20,6 +20,9 @@
 
 package com.apple.foundationdb.record.query.plan.temp;
 
+import com.apple.foundationdb.record.query.combinatorics.CrossProduct;
+import com.apple.foundationdb.record.query.combinatorics.EnumeratingIterable;
+import com.apple.foundationdb.record.query.combinatorics.EnumeratingIterator;
 import com.apple.foundationdb.record.query.plan.plans.RecordQueryPlan;
 import com.apple.foundationdb.record.query.plan.temp.Quantifier.Existential;
 import com.apple.foundationdb.record.query.plan.temp.Quantifier.ForEach;
