+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/temp/rules/RemoveSortRule.java
@@ -21,7 +21,10 @@
 package com.apple.foundationdb.record.query.plan.temp.rules;
 
 import com.apple.foundationdb.annotation.API;
+import com.apple.foundationdb.record.metadata.Index;
 import com.apple.foundationdb.record.metadata.expressions.KeyExpression;
+import com.apple.foundationdb.record.query.plan.plans.RecordQueryCoveringIndexPlan;
+import com.apple.foundationdb.record.query.plan.plans.RecordQueryIndexPlan;
 import com.apple.foundationdb.record.query.plan.plans.RecordQueryPlan;
 import com.apple.foundationdb.record.query.plan.temp.KeyPart;
 import com.apple.foundationdb.record.query.plan.temp.PlannerRule;
@@ -37,6 +40,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
+import java.util.function.Function;
 
 import static com.apple.foundationdb.record.query.plan.temp.matchers.RelationalExpressionMatchers.logicalSortExpression;
 import static com.apple.foundationdb.record.query.plan.temp.matchers.QuantifierMatchers.forEachQuantifier;
@@ -102,8 +106,21 @@ public class RemoveSortRule extends PlannerRule<LogicalSortExpression> {
                 // If we have exhausted the ordering info's keys, too, then its constituents are strictly ordered.
                 !orderingKeysIterator.hasNext() ||
                 // Also a unique index if have gone through declared fields.
-                orderingInfo.strictlyOrderedIfUnique(call.getContext()::getIndexByName, normalizedSortExpressions.size() + equalityBoundUnsorted);
+                strictlyOrderedIfUnique(innerPlan, call.getContext()::getIndexByName, normalizedSortExpressions.size() + equalityBoundUnsorted);
 
         call.yield(call.ref(strictOrdered ? innerPlan.strictlySorted() : innerPlan));
+    }
+
+    // TODO: This suggests that ordering and distinct should be tracked together.
+    public static boolean strictlyOrderedIfUnique(@Nonnull RecordQueryPlan orderedPlan, @Nonnull final Function<String, Index> getIndex, final int nkeys) {
+        if (orderedPlan instanceof RecordQueryCoveringIndexPlan) {
+            orderedPlan = ((RecordQueryCoveringIndexPlan)orderedPlan).getIndexPlan();
+        }
+        if (orderedPlan instanceof RecordQueryIndexPlan) {
+            RecordQueryIndexPlan indexPlan = (RecordQueryIndexPlan)orderedPlan;
+            Index index = getIndex.apply(indexPlan.getIndexName());
+            return index.isUnique() && nkeys >= index.getColumnSize();
+        }
+        return false;
     }
 }
