+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/query/plan/temp/TransitiveClosureTest.java
@@ -20,6 +20,7 @@
 
 package com.apple.foundationdb.record.query.plan.temp;
 
+import com.apple.foundationdb.record.query.combinatorics.TransitiveClosure;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSetMultimap;
 import com.google.common.collect.SetMultimap;
